index: use dark badge styles
Even with the backward/forward compatibility, bg-info text-dark was not sufficient contrast for WCAG 2.x AA. Forcing text-dark also had the drawback of not working well in "dark mode" themes. So if we can't use text-dark, then Bootstrap 5.0 and 5.1 will need to use the default text color. That leaves "primary", "secondary", "success", "danger" and "dark". primary and secondary are already used for buttons to Join or Cancel. success did not have sufficient contrast in all versions that we tested. danger is specifically designed to scare/alert people, so that seemed wrong. dark is the only option that has sufficient contrast, isn't designed to scare people, and does not conflict with the mental model of action buttons.