webservice: use string 'null' when recording_type is null
This allows the data to be loaded in the database and not throw an exception. In theory, the value should never be null, but somehow it is for one of our users. In the rare case where this occurs, Moodle may show the string "null", but I'm not too worried about that being the worst-case.